+++ src/RcppExports.cpp
@@ -6,21 +6,22 @@
 using namespace Rcpp;
 
 // rcpp_ngram
-List rcpp_ngram(List x, int n, String delim);
-RcppExport SEXP _textrecipes_rcpp_ngram(SEXP xSEXP, SEXP nSEXP, SEXP delimSEXP) {
+List rcpp_ngram(List x, int n, int n_min, String delim);
+RcppExport SEXP _textrecipes_rcpp_ngram(SEXP xSEXP, SEXP nSEXP, SEXP n_minSEXP, SEXP delimSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< List >::type x(xSEXP);
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    Rcpp::traits::input_parameter< int >::type n_min(n_minSEXP);
     Rcpp::traits::input_parameter< String >::type delim(delimSEXP);
-    rcpp_result_gen = Rcpp::wrap(rcpp_ngram(x, n, delim));
+    rcpp_result_gen = Rcpp::wrap(rcpp_ngram(x, n, n_min, delim));
     return rcpp_result_gen;
 END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_textrecipes_rcpp_ngram", (DL_FUNC) &_textrecipes_rcpp_ngram, 3},
+    {"_textrecipes_rcpp_ngram", (DL_FUNC) &_textrecipes_rcpp_ngram, 4},
     {NULL, NULL, 0}
 };