+++ NEWS.md
@@ -1,6 +1,7 @@
 # textrecipes (development version)
 
 * stopwords package have been moved from Imports to Suggests.
+* `step_ngram()` gained an argument `min_num_tokens` to be able to return multiple n-grams together. (#90)
 
 # textrecipes 0.2.3